+++ libs/server/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./lib/backup-eligible";
 export * from "./lib/backup-successful";
+export * from "./lib/credential";
 export * from "./lib/last-present";
 export * from "./lib/last-verified";